Major changes, albeit with no big end user changes:
Changed from vim to lua to define the neovim config. Functionality is the same,
only added a variable to neovim that gets defined before sourcing the default
(still vim, no lua) vimrc. In vimrc, added a check to see if the variable is
defined; if so, the plugins defined there won't be used. Now using packer.nvim
instead of minpac.