Changes:
- Neovim: stopped using floating window (easier to see diffs this way), and
changed command for updating treesitter parsers (minpac does not require the
':', packer.nvim does)
- Vim: new version! 9.0 introduces new vimscript (will keep old one because,
for me, Neovim is for an PDE, and vim is for everywhere I cannot get Neovim (and
there, version 9 will not be available yet). If you want to try something, do
'set wildoptions=pum' for Neovim-like commmand line popup menu. Now for the
changes: changed how the match I use for highlighting 80th character is loaded.