Changes:
- i3: just some variable changes, and changed invocation of xss-lock. The switch
'--transfer-sleep-lock' is not making a difference from what I tested.
- Vim: added 'suffixesadd' option. Example: in python files with local imports,
gf will try and add .py to the filename, going to the right file.
- Packages: finally did something I've wanted to do for a long time: group the
packages into groups and keep comments on those that are weirder. Also, added
laptop-detect, to choose if optimus-manager is installed, and igrep, which uses
ripgrep in the background to show results.