Changes:
- i3: added day of week in status bar. changed 1 (one) color for borders in
unfocused windows, the width of borders, and the alignment of the title (this is
subject to change in next commits).
- Fzf: changed options to always have quotes in zsh. Vim was picking them up,
but zsh not so much (the ones that had quotes already, particularly). Took the
chance to add conditional preview windows in fzf (just like telescope had), and
made the windows for pickers bigger.
- Tmux: added day of week in status bar.
- Vim: setting default options in fzf inside vim was useless, for basic uses
there is no need to set options. Removed autohighlighting, <leader>,
(leader-comma) will remove highlighting.
- Scripts: the packages.sh and aur.sh look like they've been changed a lot, but
I just changed from tabs to spaces in the alignment of commentary.