Changes:
- Removed default pull behaviour from gitconfig. If rebase is needed (as the
case could be with a pull when unpushed commits are already done), use the
--rebase flag
- Neovim: changed some lsp servers around, algo added some restrictions to
treesitter (I am considering disabling treesitter indentation, but we shall see)
- Vim: minor changes (removed unused commands, old comments, etc)
- Zsh: added in-house (as in already existed) alternative to zoxide (the
autojump plugin that I use): cdr <Tab> will show all the recent directories for
quick jumping
- Packages: changed onedrive from the git package to the normal

This is the last commit before the New Year (I hope). If you are reading this,
have a nice 2023!