Changes:
- Zathura: <Space> and <S-Space> now do page up and page down, instead of acting
like <C-d> and its opposite. Added keybinds for the dual page mode to change
offset.
- Git: git ac now opens the diff in the EDITOR (in vim, do :vs and have on one
side the message and in the other the diff, useful stuff). Added diff-highlight
and git-jump to the PATH so that we can use them. Added explicit no-rebase on
pull
- Zsh: minor change in one of the prompts. Changed gitconfig alias.
- Scripts: added git-extra.sh to link the git binaries that were not in PATH,
also added a whole lot of new packages (it's better to see the diff for that)